Move code to lib/servant.py
Move to lib/servant.py
- load_servants()
- load_servant()
- unplayable_servant_ids(),
- ServantLogger
